requirements.txt
Django==2.2.5
psycopg2-binary==2.8.2
pytz==2019.1
Pillow==6.0.0
django-bower==5.2.0
nodeenv==1.3.3
redis==3.2.1
django-redis==4.10.0
rq==1.0
django-rq==2.0
rq-scheduler==0.9
supervisor==4.0.3
pyvirtualdisplay==0.2.1
dbfpy==2.3.1
gsconfig==1.0.10
django-user-accounts==2.1.0
django-simple-captcha==0.5.11
django-bootstrap-form==3.4
python-magic==0.4.15
easy-thumbnails==2.6
geopandas==0.4.1
django-leaflet==0.24.0
geojson==2.5.0
